+++ spec/puppet-check/puppet_parser_spec.rb
@@ -12,7 +12,7 @@ describe PuppetParser do
     it 'puts a bad syntax Puppet manifest in the error files array' do
       PuppetParser.manifest([fixtures_dir + 'manifests/syntax.pp'], false, false, [])
       # stupid Puppet deprecation warning
-      if (Puppet::PUPPETVERSION.to_f >= 4.9) && (RUBY_VERSION.to_f < 2.1)
+      if RUBY_VERSION.to_f < 2.1
         expect(PuppetCheck.error_files[0]).to match(%r{^#{fixtures_dir}manifests/syntax.pp:\nSupport for ruby version.*\n.*\nThis Variable has no effect.*\nIllegal variable name})
       else
         expect(PuppetCheck.error_files[0]).to match(%r{^#{fixtures_dir}manifests/syntax.pp:\nThis Variable has no effect.*\nIllegal variable name})