+++ spec/spec_helper.rb
@@ -5,6 +5,7 @@ module Variables
   extend RSpec::SharedContext
 
   let(:fixtures_dir) { File.dirname(__FILE__) + '/fixtures/' }
+  let(:octocatalog_diff_dir) { File.dirname(__FILE__) + '/octocatalog-diff/' }
 end
 
 RSpec.configure do |config|