+++ spec/system/system_spec.rb
@@ -12,13 +12,13 @@ describe PuppetCheck do
 
       expect { cli }.not_to raise_exception
 
-      expect(PuppetCheck.error_files.length).to eql(9)
+      expect(PuppetCheck.error_files.length).to eql(8)
       # stupid Puppet deprecation warning
-      if (Puppet::PUPPETVERSION.to_f >= 4.9) && (RUBY_VERSION.to_f < 2.1)
-        expect(PuppetCheck.warning_files.length).to eql(11)
+      if RUBY_VERSION.to_f < 2.1
+        expect(PuppetCheck.warning_files.length).to eql(10)
         expect(PuppetCheck.clean_files.length).to eql(10)
       else
-        expect(PuppetCheck.warning_files.length).to eql(10)
+        expect(PuppetCheck.warning_files.length).to eql(9)
         expect(PuppetCheck.clean_files.length).to eql(11)
       end
       expect(PuppetCheck.ignored_files.length).to eql(1)
@@ -43,8 +43,8 @@ describe PuppetCheck do
 
       expect { tasks }.not_to raise_exception
 
-      expect(PuppetCheck.error_files.length).to eql(9)
-      expect(PuppetCheck.warning_files.length).to eql(10)
+      expect(PuppetCheck.error_files.length).to eql(8)
+      expect(PuppetCheck.warning_files.length).to eql(9)
       expect(PuppetCheck.clean_files.length).to eql(11)
       expect(PuppetCheck.ignored_files.length).to eql(1)
     end